Direct Pagefind what parts of the pages to index
We are now using Pagefind to perform a client-side search. To avoid
searching even the uninteresting parts of the pages (like navbar,
footer, etc), let's mark specifically the body of the pages that we want
to be searchable.

The front page is specifically not marked as searchable because it is
too generic to be a useful search result.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>