dark-mode: invert the images in the `About` section
These images simply have too-dark labels otherwise.

The proper solution would likely involve a separate set of images in
that section, specifically targeting dark mode, with a
specially-designed color palette. But I am not the person you want to do
that, there are experts for this kind of stuff out there in the world, I
ain't one of them.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>