js: add correct href to "Show all results..."
There was some js sourcery happening to make a click on the
"Show all results..." entry in the search results dropdown go to the
search page.

Instead of doing the sourcery on click, set a proper `href` attribute on
the `<a>`. This also makes browsers show a hand cursor when hovering the
entry in the dropdown.

Signed-off-by: Toon Claes <toon@iotcl.com>